Changed types of constant vectors to arrays and for variables to size_t

+++ src/AES/aes.cpp
@@ -15,7 +15,7 @@ void encrypt_block(vector <unsigned char>& key, unsigned char* plaintext) {
     sub_bytes(state);
     shift_rows(state);
 
-    for (int i=1; i<n_rounds; i++) {
+    for (size_t i=1; i<n_rounds; i++) {
         mix_columns(state);
         add_round_key(state, round_keys[i]);
         sub_bytes(state);
@@ -32,7 +32,7 @@ void decrypt_block(vector <unsigned char>& key, unsigned char* ciphertext) {
     std::vector <std::vector <int>> state { bytes2matrix(ciphertext) };
     add_round_key(state, round_keys[n_rounds]);
 
-    for (int i = n_rounds-1; i>0; i--) {
+    for (size_t i = n_rounds-1; i>0; i--) {
         inv_shift_rows(state);
         sub_bytes(state, inv_s_box);
         add_round_key(state, round_keys[i]);
